% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkde_fun.R
\name{bw.calc}
\alias{bw.calc}
\title{Bandwidth selection tool for gkde()}
\usage{
bw.calc(points, dist.method = "Pythagorean", boot.n = 0,
  sam.size = 100, maxram = 4)
}
\arguments{
\item{points}{A two column data frame in the form (lon,lat) or (x,y)}

\item{dist.method}{Haversine or Pythagorean distance implemented.}

\item{boot.n}{Number of bootstrap replicates to do. If set, will do bootstrap even on small datasets. Default 1000.}

\item{sam.size}{Bootstrap sample size.}

\item{maxram}{Maximum RAM, in Gigabytes.}
}
\description{
Calculate the optimal bandwidth or bootstrap the bandwidth selection for lage datasets.
}
\examples{
require(raster)
grid = raster::raster(nrows=18, ncols=36, xmn=-180, xmx=180, ymn=-90, ymx=90, vals=NULL)
grid = raster::setValues(grid,values=(as.vector(seq(1:raster::ncell(grid)))))
points = cbind(
      c(seq(xmin(grid), xmax(grid), length.out=1000), 
               seq(xmax(grid), xmin(grid), length.out=1000)), 
      c(seq(ymin(grid), ymax(grid), length.out=1000), 
               seq(ymin(grid), ymax(grid), length.out=1000))
               )
bandwidth = bw.calc(points);
}
