% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkde_fun.R
\name{dist2point}
\alias{dist2point}
\title{Calculate distance to points surfaces}
\usage{
dist2point(grid, points, parallel = TRUE, nclus = 4,
  dist.method = "Haversine", maxram = 4)
}
\arguments{
\item{grid}{A raster object to match.}

\item{points}{A two column data frame in the form (lon,lat) or (x,y)}

\item{parallel}{TRUE or FALSE, should this code be executed in parallel.}

\item{nclus}{IF parallel==TRUE then how many cores in the cluster.}

\item{dist.method}{Which distance should we use? Haversine for lat/long projections,or Pythagorean for flat images and/or small areas.}

\item{maxram}{Maximum theoretical RAM usage. Will be divided by nclus for parallel jobs.}
}
\description{
This function calculates the minimum distance to points raster.
}
\examples{
require(raster)
grid = raster::raster(nrows=81, ncols=162, xmn=-180, xmx=180, ymn=-90, ymx=90, vals=NULL)
grid = raster::setValues(grid,values=(as.vector(seq(1:raster::ncell(grid)))))
points = cbind(
      c(seq(xmin(grid), xmax(grid), length.out=1000), 
               seq(xmax(grid), xmin(grid), length.out=1000)), 
      c(seq(ymin(grid), ymax(grid), length.out=100), 
               seq(ymin(grid), ymax(grid), length.out=100))
               )
plot(grid); points(points);
di = dist2point(grid, points, parallel=TRUE, maxram = 0.5, nclus = 4, dist.method='Pythagorean')
plot(di, col = viridis::viridis(9))
points(points)
}
