% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkde_fun.R
\name{gkde}
\alias{gkde}
\title{Geographic Kernel Density Estimator using linear or Haversine distances}
\usage{
gkde(grid, points, parallel = TRUE, nclus = 4, dist.method = "Haversine")
}
\arguments{
\item{grid}{A raster object to match.}

\item{points}{A two column data frame in the form (lon,lat) or (x,y)}

\item{parallel}{TRUE or FALSE, should this code be executed in parallel.}

\item{nclus}{IF parallel==TRUE then how many cores in the cluster.}

\item{dist.method}{Which distance should we use? Haversine for lat/long projections, 
or Pythagorean for flat images and/or small areas.}
}
\description{
This function requests data from a private service that
mirrors GBIF distribution data at cloud.diversityoflife.org. Account required.
}
\examples{
 \dontrun {
}
}
